Siguen las pruebas de los indices secundarios. Se implementó con Nico el método siguienteBloque() en BloqueManager para poder ir obteniendo todos los bloques de datos cuando la operación es un SELECT* (sin ordenar).

+++ Common/Clave/Clave.cpp
@@ -595,7 +595,7 @@
 	{
 		fecha* clave = (static_cast<fecha*>(this->getValor()));
 
-		salida<<"Clave: "<<clave->dia << "/" << clave->mes << "/" << clave->anio;
+		salida<<"Clave: "<< (short)clave->dia << "/" << (short)clave->mes << "/" << clave->anio;
 		salida<<" Referencia: "<<this->getReferencia()<<endl;
 	}
